href

+++ gisportal/beheer/geo-package.php
@@ -25,7 +25,7 @@ $basicPage->writeLog('geo-package.php 1');
 			if ($g['id']>=1) {
 				$db->delete('geopackages','id='.$g['id']);
 				$delfs=glob($basicPage->getConfig('geo-mappen').'/geo-packages/gpid-'.$g['id'].'/*.*');
-				id ($delfs) foreach ($delfs as $delf) {
+				if ($delfs) foreach ($delfs as $delf) {
 					unlink($delf);
 				}
 				$fname=$basicPage->getConfig('geo-mappen').'/geo-packages/gpid-'.$g['id'];
@@ -37,7 +37,6 @@ $basicPage->writeLog('geo-package.php 1');
 			}
 		} else {
 			if ($g) {
-	/*
 				if ($func=='opslaan' || $func=='cache-legen') {
 					if ($func=='cache-legen') {
 						if (file_exists('indata.data')) {unlink('indata.data');}
@@ -299,8 +298,6 @@ $basicPage->writeLog('geo-package.php 1');
 				if ($tab2) {$r.=$tab2;}
 				$r.='</div>';
 				$basicPage->add_js_ready('$( "#tabs" ).tabs({heightStyle: \'auto\'});');
-*/
-				
 			} else {
 				$basicPage->fout('Internal error','Geopackage niet gevonden.');
 			}
